# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Study-ing API
  version: 0.2.0
  description: |
    Study-ing 프론트엔드와 Spring Boot 백엔드가 공유하는 API 계약입니다.

    구현 원칙:
    - GitLab access token은 브라우저에 반환하지 않습니다.
    - 인증된 API는 HttpOnly 세션 쿠키를 사용합니다.
    - Workspace ID로 연결 프로젝트를 결정하며 요청에서 임의 프로젝트를 받지 않습니다.
    - 날짜는 `YYYY-MM-DD`, 시각은 timezone을 포함한 ISO 8601 문자열입니다.
    - 수정 요청은 `expectedRevision` 또는 `lastCommitId`로 낙관적 잠금을 적용합니다.
  contact:
    name: Study-ing Team
servers:
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 서버
...
paths:
  /auth/gitlab/login:
    get:
      tags: [Auth]
      summary: GitLab OAuth 로그인 시작
      operationId: startGitLabLogin
      x-owner: member-1
      x-implementation-status: implemented
      security: []
      parameters:
        - name: returnUrl
          in: query
          required: false
          schema:
            type: string
            default: /
            pattern: ^/
          description: 로그인 완료 후 이동할 앱 내부 경로
      responses:
        "302":
          description: GitLab authorize URL로 이동
          headers:
            Location:
              schema:
                type: string
                format: uri
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /auth/gitlab/callback:
    get:
      tags: [Auth]
      summary: GitLab OAuth callback 검증과 프론트 전환 화면 이동
      operationId: stageGitLabLogin
      x-owner: member-1
      x-implementation-status: implemented
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: 승인 코드를 서버 세션에 일회성으로 보관하고 프론트 `/auth/callback`으로 이동
          headers:
            Location:
              schema:
                type: string
  /auth/gitlab/complete:
    post:
      tags: [Auth]
      summary: 대기 중인 GitLab OAuth 승인 완료
      description: 서버 세션의 일회성 승인 코드를 소비해 token 교환, 사용자 저장, 세션 ID 회전을 완료합니다.
      operationId: completeGitLabLogin
      x-owner: member-1
      x-implementation-status: implemented
      parameters:
        - name: X-CSRF-TOKEN
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 로그인 완료 후 이동할 앱 내부 경로
          content:
            application/json:
              schema:
                type: object
                required: [returnUrl]
                properties:
                  returnUrl:
                    type: string
                    pattern: ^/
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /auth/me:
    get:
      tags: [Auth]
      summary: 현재 로그인 사용자 조회
      operationId: getCurrentUser
      x-owner: member-1
      x-implementation-status: implemented
      responses:
        "200":
          description: 현재 GitLab 로그인 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "401":
          description: 로그인 세션 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthenticatedResponse"
...
  /auth/profile:
    put:
      tags: [Auth]
      summary: 가입 프로필 저장 또는 수정
      operationId: updateAccountProfile
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountProfileRequest"
      responses:
        "200":
          description: 저장된 서비스 프로필
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedGitLabUser"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /auth/preferences:
    patch:
      tags: [Auth]
      summary: 개인 화면 테마 설정 저장
      operationId: updateAccountPreferences
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountPreferencesRequest"
      responses:
        "200":
          description: 변경된 사용자 프로필과 테마 설정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedGitLabUser"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /auth/account:
    delete:
      tags: [Auth]
      summary: 계정과 암호화 OAuth credential 삭제
      description: 활성 Workspace의 Owner는 먼저 해당 Workspace를 삭제해야 합니다. GitLab의 기존 파일과 commit은 삭제하지 않습니다.
      operationId: deleteAccount
      x-implementation-status: implemented
      responses:
        "204":
          description: 계정 삭제와 세션 종료 완료
        "409":
          $ref: "#/components/responses/Conflict"
...
  /workspaces/{workspaceId}/sessions/{date}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/SessionDate"
    get:
      tags: [Sessions]
      summary: 특정 날짜 일정
      operationId: getSession
      x-owner: member-2
      x-implementation-status: implemented
      responses:
        "200":
          description: 일정 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Sessions]
      summary: revision과 lastCommitId를 확인하고 일정 수정
      operationId: updateSession
      x-owner: member-2
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: 수정된 일정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Sessions]
      summary: session.yml을 삭제하지 않고 cancelled 처리
      operationId: cancelSession
      x-owner: member-2
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedRevision, lastCommitId, changeReason]
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                lastCommitId:
                  type: string
                changeReason:
                  type: string
                  minLength: 1
                  maxLength: 300
      responses:
        "200":
          description: cancelled 상태의 일정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionDate:
      name: date
      in: path
      required: true
      schema:
        type: string
        format: date
...
  responses:
    InvalidRequest:
      description: 요청 형식 또는 비즈니스 입력값 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: INVALID_REQUEST
            message: 요청값을 확인해 주세요.
            details:
              field: title
            requestId: req_01J
            timestamp: "2026-07-25T12:00:00Z"
...
    NotFound:
      description: Workspace, Session, Item 또는 파일을 찾지 못함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: RESOURCE_NOT_FOUND
            message: 요청한 리소스를 찾지 못했습니다.
            timestamp: "2026-07-25T12:00:00Z"
    Conflict:
      description: revision, lastCommitId 또는 중복 리소스 충돌
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: SESSION_REVISION_CONFLICT
            message: 다른 사용자가 먼저 수정했습니다. 최신 내용을 확인해 주세요.
            details:
              expectedRevision: 3
              actualRevision: 4
            timestamp: "2026-07-25T12:00:00Z"
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: 프론트 분기와 테스트에서 사용하는 안정적인 오류 코드
        message:
          type: string
          description: 사용자에게 표시 가능한 한국어 메시지
        details:
          type: object
          additionalProperties: true
          description: 필드 오류나 충돌 최신값
        requestId:
          type: string
          description: 로그 추적 ID
        timestamp:
          type: string
          format: date-time
...
    CurrentUserResponse:
      type: object
      required: [authenticated, mode, user]
      properties:
        authenticated:
          const: true
        mode:
          const: gitlab-oauth
        user:
          $ref: "#/components/schemas/AuthenticatedGitLabUser"
    UnauthenticatedResponse:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          const: false
    AuthenticatedGitLabUser:
      type: object
      required: [id, username, name, profileCompleted, timezone, themeMode, accentColor]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        avatarUrl:
          type: [string, "null"]
          format: uri
        webUrl:
          type: [string, "null"]
          format: uri
        profileCompleted:
          type: boolean
        repositoryFileName:
          type: [string, "null"]
        timezone:
          type: string
        termsVersion:
          type: [string, "null"]
        termsAgreedAt:
          type: [string, "null"]
          format: date-time
        privacyVersion:
          type: [string, "null"]
        privacyAgreedAt:
          type: [string, "null"]
          format: date-time
        minimumAgeConfirmedAt:
          type: [string, "null"]
          format: date-time
        requiresReconsent:
          type: boolean
        themeMode:
          type: string
          enum: [LIGHT, DARK]
        accentColor:
          type: string
          enum: [PURPLE, BLUE, TEAL, ORANGE, ROSE]
    UpdateAccountProfileRequest:
      type: object
      required: [displayName, repositoryFileName, timezone, acceptTerms, acceptPrivacy, confirmMinimumAge]
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 40
        repositoryFileName:
          type: string
          maxLength: 80
        timezone:
          type: string
        acceptTerms:
          type: boolean
          description: 최초 onboarding 또는 재동의 시 이용약관 동의
        acceptPrivacy:
          type: boolean
          description: 최초 onboarding 또는 재동의 시 개인정보 처리 안내 동의
        confirmMinimumAge:
          type: boolean
          description: 만 14세 이상 확인
    UpdateAccountPreferencesRequest:
      type: object
      required: [themeMode, accentColor]
      properties:
        themeMode:
          type: string
          enum: [LIGHT, DARK]
        accentColor:
          type: string
          enum: [PURPLE, BLUE, TEAL, ORANGE, ROSE]
...
    SessionType:
      type: string
      enum: [algorithm, english, cs, free]
    SessionStatus:
      type: string
      enum: [active, cancelled]
    SubmitType:
      type: string
      enum: [link, text, code, mixed]
    StudyItem:
      type: object
      required: [id, order, title, type, submitType, required, status]
      properties:
        id:
          type: string
          pattern: ^item-[a-f0-9]{6,16}$
        order:
          type: integer
          minimum: 1
        title:
          type: string
        type:
          $ref: "#/components/schemas/SessionType"
        source:
          type: [string, "null"]
        url:
          type: [string, "null"]
          format: uri
        submitType:
          $ref: "#/components/schemas/SubmitType"
        required:
          type: boolean
        status:
          type: string
          enum: [active, cancelled]
    StudySession:
      type: object
      required:
        [date, folder, revision, type, title, description, status, deadline, items, lastCommitId]
      properties:
        date:
          type: string
          format: date
        folder:
          type: string
          pattern: ^\d{6}$
        revision:
          type: integer
          minimum: 1
        type:
          $ref: "#/components/schemas/SessionType"
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        deadline:
          type: string
          format: date-time
        secondaryDeadline:
          type: [string, "null"]
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/StudyItem"
        lastCommitId:
          type: string
    SessionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/StudySession"
    SessionItemDraft:
      type: object
      required: [title, type, submitType, required]
      properties:
        id:
          type: string
          description: 수정 시 기존 항목 유지에만 전달
        title:
          type: string
          minLength: 1
          maxLength: 120
        type:
          $ref: "#/components/schemas/SessionType"
        source:
          type: string
          maxLength: 80
        url:
          type: string
          format: uri
        submitType:
          $ref: "#/components/schemas/SubmitType"
        required:
          type: boolean
    CreateSessionRequest:
      type: object
      required: [date, type, title, description, deadline, items]
      properties:
        date:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/SessionType"
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        deadline:
          type: string
          format: date-time
        secondaryDeadline:
          type: [string, "null"]
          format: date-time
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SessionItemDraft"
    UpdateSessionRequest:
      allOf:
        - $ref: "#/components/schemas/CreateSessionRequest"
        - type: object
          required: [expectedRevision, lastCommitId, changeReason]
          properties:
            expectedRevision:
              type: integer
              minimum: 1
            lastCommitId:
              type: string
            changeReason:
              type: string
              minLength: 1
              maxLength: 300
...
security:
  - sessionCookie: []